Improved History screen structure and synchronization with location statuses
