+++ app/src/main/java/com/example/myapplication/ui/screens/moderator/nav/ContentModerator.kt
@@ -16,7 +16,7 @@ import com.example.myapplication.viewmodel.UsersViewModel
 
 
 @Composable
-fun ContentModerator(padding: PaddingValues , navController: NavHostController) {
+fun ContentModerator(padding: PaddingValues, navController: NavHostController) {
 
     val usersViewModel: UsersViewModel = viewModel()
     val placesViewModel: PlacesViewModel = viewModel()
@@ -27,23 +27,21 @@ fun ContentModerator(padding: PaddingValues , navController: NavHostController) 
         navController = navController,
         startDestination = RouteTab.Dashboard
 
-    ){
+    ) {
 
 
         composable<RouteTab.Dashboard> {
-            Dashboard(placesViewModel= placesViewModel, usersViewModel = usersViewModel)
+            Dashboard(placesViewModel = placesViewModel, usersViewModel = usersViewModel)
         }
 
         composable<RouteTab.History> {
-            History()
+            History(viewModel = placesViewModel)
         }
 
 
         composable<RouteTab.Profile> {
             Profile(usersViewModel = usersViewModel)
         }
-
-
 
 
     }
